refactor: move Configuration unit test to inner mods
