+++ packages/configuration/src/lib.rs
@@ -157,136 +157,3 @@ impl From<figment::Error> for Error {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::Configuration;
-
-    #[cfg(test)]
-    fn default_config_toml() -> String {
-        let config = r#"log_level = "info"
-                                mode = "public"
-                                db_driver = "Sqlite3"
-                                db_path = "./storage/tracker/lib/database/sqlite3.db"
-                                announce_interval = 120
-                                min_announce_interval = 120
-                                on_reverse_proxy = false
-                                external_ip = "0.0.0.0"
-                                tracker_usage_statistics = true
-                                persistent_torrent_completed_stat = false
-                                max_peer_timeout = 900
-                                inactive_peer_cleanup_interval = 600
-                                remove_peerless_torrents = true
-
-                                [[udp_trackers]]
-                                enabled = false
-                                bind_address = "0.0.0.0:6969"
-
-                                [[http_trackers]]
-                                enabled = false
-                                bind_address = "0.0.0.0:7070"
-                                ssl_enabled = false
-                                ssl_cert_path = ""
-                                ssl_key_path = ""
-
-                                [http_api]
-                                enabled = true
-                                bind_address = "127.0.0.1:1212"
-                                ssl_enabled = false
-                                ssl_cert_path = ""
-                                ssl_key_path = ""
-
-                                [http_api.access_tokens]
-                                admin = "MyAccessToken"
-
-                                [health_check_api]
-                                bind_address = "127.0.0.1:1313"
-        "#
-        .lines()
-        .map(str::trim_start)
-        .collect::<Vec<&str>>()
-        .join("\n");
-        config
-    }
-
-    #[test]
-    fn configuration_should_have_default_values() {
-        let configuration = Configuration::default();
-
-        let toml = toml::to_string(&configuration).expect("Could not encode TOML value");
-
-        assert_eq!(toml, default_config_toml());
-    }
-
-    #[test]
-    fn configuration_should_contain_the_external_ip() {
-        let configuration = Configuration::default();
-
-        assert_eq!(configuration.external_ip, Some(String::from("0.0.0.0")));
-    }
-
-    #[test]
-    fn configuration_should_be_saved_in_a_toml_config_file() {
-        use std::{env, fs};
-
-        use uuid::Uuid;
-
-        // Build temp config file path
-        let temp_directory = env::temp_dir();
-        let temp_file = temp_directory.join(format!("test_config_{}.toml", Uuid::new_v4()));
-
-        // Convert to argument type for Configuration::save_to_file
-        let config_file_path = temp_file;
-        let path = config_file_path.to_string_lossy().to_string();
-
-        let default_configuration = Configuration::default();
-
-        default_configuration
-            .save_to_file(&path)
-            .expect("Could not save configuration to file");
-
-        let contents = fs::read_to_string(&path).expect("Something went wrong reading the file");
-
-        assert_eq!(contents, default_config_toml());
-    }
-
-    #[cfg(test)]
-    fn create_temp_config_file_with_default_config() -> String {
-        use std::env;
-        use std::fs::File;
-        use std::io::Write;
-
-        use uuid::Uuid;
-
-        // Build temp config file path
-        let temp_directory = env::temp_dir();
-        let temp_file = temp_directory.join(format!("test_config_{}.toml", Uuid::new_v4()));
-
-        // Convert to argument type for Configuration::load_from_file
-        let config_file_path = temp_file.clone();
-        let path = config_file_path.to_string_lossy().to_string();
-
-        // Write file contents
-        let mut file = File::create(temp_file).unwrap();
-        writeln!(&mut file, "{}", default_config_toml()).unwrap();
-
-        path
-    }
-
-    #[test]
-    fn configuration_should_be_loaded_from_a_toml_config_file() {
-        let config_file_path = create_temp_config_file_with_default_config();
-
-        let configuration = Configuration::load_from_file(&config_file_path).expect("Could not load configuration from file");
-
-        assert_eq!(configuration, Configuration::default());
-    }
-
-    #[test]
-    fn http_api_configuration_should_check_if_it_contains_a_token() {
-        let configuration = Configuration::default();
-
-        assert!(configuration.http_api.access_tokens.values().any(|t| t == "MyAccessToken"));
-        assert!(!configuration.http_api.access_tokens.values().any(|t| t == "NonExistingToken"));
-    }
-}
